requirements.txt
pdfplumber==0.11.8
qdrant-client==1.16.2
sentence-transformers==5.2.0
requests==2.32.5
numpy==2.4.0
fastapi==0.115.6
uvicorn==0.34.0
python-dotenv==1.0.1
pydantic==2.10.4
